% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetExperimentInfo.R
\name{GetExperimentInfo}
\alias{GetExperimentInfo}
\title{Experiment information}
\usage{
GetExperimentInfo(experiment.id)
}
\arguments{
\item{experiment.id}{valid experiment identifier}
}
\value{
data frame of metadata information for an experiment
}
\description{
Query Optimizely  to read metadata information for a single experiment
}
\examples{
\dontrun{
experiment.info <- GetExperimentInfo('1234545')
#use this to extract information on a list of experiments
do.call(rbind, lapply(c(1231231,14335333'), GetExperimentInfo))
}
}
\seealso{
\code{\link{GetExperimentList}}
}

