% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetExperimentList.R
\name{GetExperimentList}
\alias{GetExperimentList}
\title{Query Optimizely  to list experiments in projects}
\usage{
GetExperimentList(project.list)
}
\arguments{
\item{project.list}{List of project identifiers}
}
\value{
data frame of all experiments in list of projects provided
}
\description{
\code{GetExperimentList} extracts the experiments and summary information in a list of projects
}
\examples{
\dontrun{
# list experiments in a project
#
experiment.list<-GetExperimentList( c('1234567','7654321') )
}
}
\seealso{
\code{\link{GetExperimentInfo}}
}

