% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetProjectList.R
\name{GetProjectList}
\alias{GetProjectList}
\title{Query Optimizely API to list projects in an account}
\usage{
GetProjectList()
}
\value{
data frame containing the response from Optimizely API
}
\description{
\code{GetProjectList} Returns a data frame of projects listed in account authorized by the token
Obtain  API token from \url{https://app.optimizely.com/tokens}
}
\examples{
\dontrun{
 # use set_token(Token) to assign your API token
project.df <- GetProjectList()
}
}
\seealso{
\code{\link{GetProjectInfo}}
}

