% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{set_token}
\alias{set_token}
\title{API token Authorizations}
\usage{
set_token(token)
}
\arguments{
\item{token}{copy token string from 'Manage Tokens'}
}
\description{
All API request needs to be authenticated. To authenticate, use token generated from
\url{https://app.optimizely.com/tokens}
}
\examples{
set_token('abcdefghijklmnopqrstuvwxyz:123456')
}

